feat: add backlight for caps lock
