+++ js/functions/add-backlight.js
@@ -1,0 +1,5 @@
+export default function addBacklight(span, btn, e) {
+  if (e.code === btn.dataset.code) {
+    btn.append(span);
+  }
+};
