+++ js/functions/create-backlight-span.js
@@ -1,0 +1,8 @@
+import createElement from './create-element.js';
+
+export default function createBacklightSpan() {
+  const span = createElement('span', 'key__backlight');
+  span.style.animationName = 'clickButton';
+
+  return span;
+}
