+++ js/functions/remove-backlight.js
@@ -1,0 +1,3 @@
+export default function removeBacklight(span) {
+  setTimeout((span) => {span.remove()}, 500, span);
+};
